chore(web): extract `getWeekEventsSlice.actions.request`dispatch call into dedicated helper function
Since this dispatch call is reused in 2 different places, this helps in keeping the logic consistent.
