+++ packages/web/src/views/Calendar/hooks/useWeek.ts
@@ -1,11 +1,10 @@
 import { Dayjs } from "dayjs";
 import { useEffect, useMemo, useState } from "react";
 import { YEAR_MONTH_DAY_FORMAT } from "@core/constants/date.constants";
-import { toUTCOffset } from "@web/common/utils/web.date.util";
+import { handleDispatchGetWeekEvents } from "@web/common/utils/event.util";
 import { Week_AsyncStateContextReason } from "@web/ducks/events/context/week.context";
 import { getSomedayEventsSlice } from "@web/ducks/events/slices/someday.slice";
 import { updateDates } from "@web/ducks/events/slices/view.slice";
-import { getWeekEventsSlice } from "@web/ducks/events/slices/week.slice";
 import { useAppDispatch } from "@web/store/store.hooks";
 import { Category_View } from "@web/views/Calendar/calendarView.types";
 
@@ -29,13 +28,11 @@ export const useWeek = (today: Dayjs) => {
 
   useEffect(() => {
     dispatch(
-      getWeekEventsSlice.actions.request({
-        startDate: toUTCOffset(start),
-        endDate: toUTCOffset(end),
-        __context: {
-          reason: Week_AsyncStateContextReason.WEEK_VIEW_CHANGE,
-        },
-      }),
+      handleDispatchGetWeekEvents(
+        start,
+        end,
+        Week_AsyncStateContextReason.WEEK_VIEW_CHANGE,
+      ),
     );
   }, [dispatch, end, start]);
